Cast `iv_length` to an int so the comparison can succeed.
The comparison of `iv_length` and `SHA256_DIGEST_LENGTH` will fail even
in a success scenario because `iv_length` is an int and
`SHA256_DIGEST_LENGTH` is an `unsigned int`. This caused the call to the
`run` method on the `OpenSSLSHA256` kernel to fail with a SystemError
since the function simply returns `NULL` instead of creating/raising an
Exception.

+++ timelock/kernels/sha256module.c
@@ -25,7 +25,7 @@ sha256_run(PyObject *self, PyObject *args)
     if (!PyArg_ParseTuple(args, "s#K", &iv, &iv_length, &n))
         return NULL;
 
-    if (iv_length != SHA256_DIGEST_LENGTH)
+    if (((int) iv_length) != SHA256_DIGEST_LENGTH)
         return NULL;
 
     memcpy(midstate, iv, SHA256_DIGEST_LENGTH);
